Split satellite_view.rs into satellite_view/ subdirectory
Split crates/rendering/src/satellite_view.rs (719 lines) into focused modules
under satellite_view/ for better maintainability.
